Settled on directory structure and fixed ESlint
