--- App.js
@@ -1,62 +1,0 @@
-import React from 'react';
-import { StatusBar } from 'expo-status-bar';
-import {
-  StyleSheet, View, Text,
-} from 'react-native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { NavigationContainer } from '@react-navigation/native';
-import Icon from 'react-native-vector-icons/Ionicons';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'flex-end',
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-  },
-  container__button: {
-    backgroundColor: '#000',
-  },
-});
-
-function Home() {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>Home!</Text>
-    </View>
-  );
-}
-
-const screenOptions = ({ route }) => ({
-  headerShown: false,
-  tabBarIcon: () => (
-    <Icon name={
-      route.name === 'Home' ? 'list-outline' : 'settings-outline'
-    }
-    />
-  ),
-});
-
-function Settings() {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>Settings!</Text>
-    </View>
-  );
-}
-
-const Tab = createBottomTabNavigator();
-
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <NavigationContainer>
-        <Tab.Navigator screenOptions={screenOptions}>
-          <Tab.Screen name="Home" component={Home} />
-          <Tab.Screen name="Settings" component={Settings} />
-        </Tab.Navigator>
-      </NavigationContainer>
-      <StatusBar />
-    </View>
-  );
-}
