+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ic-tech-4901",
-  "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "version": "0.0.0",
+  "main": "src/components/App.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
