+++ src/components/Home.js
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native';
+import React from 'react';
+
+export default function Home() {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Home!</Text>
+    </View>
+  );
+}
