+++ src/components/Settings.js
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native';
+import React from 'react';
+
+export default function Settings() {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Settings!</Text>
+    </View>
+  );
+}
